Fix: SkillParser & RegionSkills (#378)
* Fix: SkillParser & RegionSkills

- Now all skills are being parse correctly.
- Adding more skill type checks.
- Adding GM skills type.
- Now Skill Region are being cast correctly. (WIP)
- Correct implementation of skill Levels.
- Minor other fixes.

* Fix suggestions

* Fix playercommands

+++ MapleServer2/Packets/SkillBookTreePacket.cs
@@ -28,9 +28,9 @@ namespace MapleServer2.Packets
                 pWriter.WriteLong(skillTab.TabId);
                 pWriter.WriteUnicodeString(skillTab.Name);
 
-                Dictionary<int, int> skills = skillTab.SkillLevels.Where(x => x.Value > 0).ToDictionary(x => x.Key, x => x.Value);
+                Dictionary<int, short> skills = skillTab.SkillLevels.Where(x => x.Value > 0).ToDictionary(x => x.Key, x => x.Value);
                 pWriter.WriteInt(skills.Count);
-                foreach (KeyValuePair<int, int> p in skills)
+                foreach (KeyValuePair<int, short> p in skills)
                 {
                     pWriter.WriteInt(p.Key);
                     pWriter.WriteInt(p.Value);
